Create NO;U
the deal is this. earth has been invaded by shapeshifting lizards that work to sell the souls of mankind into total servitude to phyrexia, the source of all things we hate like the cold iron that freezez minds of mortal in order for us to be in a be a perfect cuisine of rape and filth... do not be alarmed howerver.